extract run_ping to new module
