+++ test/quoil_test.exs
@@ -1,7 +1,4 @@
 defmodule QuoilTest do
   use ExUnit.Case
 
-  test "the truth" do
-    assert 1 + 1 == 2
-  end
 end
